Renamed to remove java keyword
