+++ examples/minimal-java-native/services/settings.gradle
@@ -1,3 +1,3 @@
 
-rootProject.name="minimal-java-native"
+rootProject.name="minimal-javaNative"
