+++ examples/minimal-java-native/services/src/main/java/minimal/javaNative/FunctionLambdaRuntime.java
@@ -1,4 +1,4 @@
-package minimal.java.native;
+package minimal.javaNative;
 import com.amazonaws.services.lambda.runtime.events.APIGatewayProxyRequestEvent;
 import com.amazonaws.services.lambda.runtime.events.APIGatewayProxyResponseEvent;
 import io.micronaut.function.aws.runtime.AbstractMicronautLambdaRuntime;
