+++ examples/minimal-java-native/services/src/test/java/minimal/javaNative/FunctionRequestHandlerTest.java
@@ -1,4 +1,4 @@
-package minimal.java.native;
+package minimal.javaNative;
 import com.amazonaws.services.lambda.runtime.events.APIGatewayProxyRequestEvent;
 import com.amazonaws.services.lambda.runtime.events.APIGatewayProxyResponseEvent;
 import org.junit.jupiter.api.AfterAll;
